Fixed collection updater
