+++ ci/travis/script.sh
@@ -1,2 +1,2 @@
  phpunit
- ./vendor/bin/phpbench run --progress=dots --report=aggregate --report="extends: aggregate, break: [ benchmark ]"
+ ./vendor/bin/phpbench run --progress=dots --report="extends: aggregate, break: [ benchmark ]"
