+++ src/Standard/Storage/ReferenceType.php
@@ -13,7 +13,9 @@ final class ReferenceType implements TypeInterface
 {
     public function configureOptions(OptionsResolver $resolver)
     {
-        $resolver->setDefaults('class', null);
+        $resolver->setDefaults([
+            'class' => null,
+        ]);
         $resolver->setAllowedTypes('class', ['null', 'array']);
     }
 }
